requirements.txt
opencv-python
numpy
mediapipe
tqdm
scikit-learn
tensorflow
transformers
torch
tf-keras
python-dotenv
seaborn